% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSR.R
\name{FSR}
\alias{FSR}
\title{Factor Score Regression approach}
\usage{
FSR(lavModel_Analysis, data, FSmethod = "SL", data_transformations = NULL)
}
\arguments{
\item{lavModel_Analysis}{the lavModel_Analysis object}

\item{data}{set to fit}

\item{FSmethod}{Method to be used to extract factor scores. Default to \code{"SL"} for the Skrondal and Laake approach that uses regression (\code{"regression"}) factor scores for the independendent variables and \code{"Bartlett"} factor scores for the dependent variables.}

\item{data_transformations}{Data transformations}
}
\description{
Factor Score Regression approach
}
