% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMS.R
\name{LMS}
\alias{LMS}
\title{Latent moderated strctured equations by Klein and Moosbrugger (2000), the ML approach to nonlinear SEM}
\usage{
LMS(
  lavModel_Analysis,
  data,
  data_transformations = NULL,
  prefix = 1,
  algorithm = "INTEGRATION"
)
}
\arguments{
\item{lavModel_Analysis}{the lavModel_Analysis object}

\item{data}{set to fit}

\item{prefix}{an arbitrary prefix for the data set. This prevents issues when using parallelization and Mplus.}

\item{algorithm}{algorithm to use. Default to INTEGRATION.}
}
\description{
Latent moderated strctured equations by Klein and Moosbrugger (2000), the ML approach to nonlinear SEM
}
