% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.powerNLSEM.R
\name{plot.powerNLSEM}
\alias{plot.powerNLSEM}
\title{plot powerNLSEM object}
\usage{
\method{plot}{powerNLSEM}(
  x,
  min_num_bins = 10,
  plot = "power_model",
  power_modeling_method = NULL,
  se = FALSE,
  power_aim = NULL,
  alpha = NULL,
  ...
)
}
\arguments{
\item{x}{object of class powerNLSEM}

\item{min_num_bins}{minimal number of bins used for aggregating results. Default to 10.}

\item{plot}{Character indicating what type of plot to create. Default to "power_model", referencing to the prediction of significant parameters using the model specified in power_modeling_method.}

\item{power_modeling_method}{Character indicating the power modeling method used. This is only relevant when \code{plot = "power_model"} is used. Default to \code{NULL}, indicating to use the same power modeling method as was used in the \code{powerNLSEM} function.}

\item{se}{Logical indicating to use confidence intervals based on normal approximation using the standard errors. Default to \code{FALSE}.}

\item{power_aim}{Power level to be included into the plot with respective N. If \code{NULL} the same power level as in the \code{powerNLSEM} function will be used. If set to \code{0} no power level and corresponding N will be plotted. Default to \code{NULL}, indicating to use the same power modeling method as was used in the \code{powerNLSEM} function.}

\item{alpha}{Alpha value used for confidence intervals, when \code{se = TRUE}. Default to \code{NULL}, indicating to use the same alpha as was used in the powerNLSEM function. This does not influence the significance decision, although same alpha is used per default.}

\item{...}{Additional arguments passed on to the plot function.}
}
\value{
Returns \code{ggplot} object of the type specified in plot.
}
\description{
plot powerNLSEM object
}
