% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_search.R
\name{power_search}
\alias{power_search}
\title{Search function to find N for desired power}
\usage{
power_search(
  POI,
  method,
  lavModel,
  lavModel_Analysis,
  data_transformations,
  search_method,
  power_modeling_method,
  R = 1000,
  power_aim = 0.8,
  alpha = 0.05,
  CORES,
  verbose,
  Ns = NULL,
  N_start = nrow(lavModel) * 10,
  type = "u",
  steps = 10,
  lb = nrow(lavModel),
  switchStep = round(steps/2),
  uncertainty_method = "",
  FSmethod = "SR",
  seeds
)
}
\arguments{
\item{POI}{Parameter Of Interest as a vector of strings. Must be in lavaan-syntax without any spaces. Nonlinear effects should have the same ordering as in model.}

\item{method}{Method used to fit to the data. Can be LMS or UPI.}

\item{lavModel}{lavModel object describing the model.}

\item{lavModel_Analysis}{lavModel object containg the parameters to be estimated.}

\item{data_transformations}{Object containing info on data transformations.}

\item{search_method}{String stating the search method. Default to \code{"smart"}. Alternative is \code{"bruteforce"}.}

\item{power_modeling_method}{Power modeling method used to model significant parameter estimates. Default to \code{"probit"} indicating glm with probit link function with sqrt(n) as predictor. Alternative is \code{"logit"}.}

\item{R}{Total number of models to be fitted. Higher number results in higher precision and longer runtime.}

\item{power_aim}{Minimal power value to approximate. Default to .8.}

\item{alpha}{Type I-error rate. Default to .05.}

\item{CORES}{Number of cores used for parallelization. Default to number of available cores - 2.}

\item{verbose}{Logical whether progress should be printed in console. Default to TRUE.}

\item{Ns}{Sample sizes used in power estimation process. Default to \code{NULL}.}

\item{N_start}{Starting sample size. Default to \code{nrow(lavModel)*10}}

\item{type}{Default to \code{"u"}.}

\item{steps}{Steps used in search_method = "smart", i.e., the smart algorithm. This is ignored if bruteforce is used. Default to 10.}

\item{lb}{Lower bound of N used in search. Default to \code{nrow(lavModel)}}

\item{switchStep}{Steps after which smart search method changes from exploration to exploitation. Default to \code{round(steps/2)}}

\item{uncertainty_method}{Uncertainty method used for confidence intervals. Default to \code{""}}

\item{FSmethod}{Method to be used to extract factor scores. Default to \code{"SL"} for the Skrondal and Laake approach that uses regression (\code{"regression"}) factor scores for the independendent variables and \code{"Bartlett"} factor scores for the dependent variables.}

\item{seeds}{Seeds for reproducibility.}
}
\description{
Search function to find N for desired power
}
