% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.powerNLSEM.R
\name{print.summary.powerNLSEM}
\alias{print.summary.powerNLSEM}
\title{print summary for powerNLSEM objects}
\usage{
\method{print}{summary.powerNLSEM}(x, ...)
}
\arguments{
\item{x}{Prints the summary of a \code{powerNLSEM} object. \code{x} must be of class \code{"summary.powerNLSEM"}.}

\item{...}{Further arguments to use in \code{print}.}
}
\value{
Prints output of summmary of \code{powerNLSEM} object into the console (objects of class \code{summary.powerNLSEM}), but does not change object itself.
}
\description{
print summary for powerNLSEM objects
}
