% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalyse.powerNLSEM.R
\name{reanalyse.powerNLSEM}
\alias{reanalyse.powerNLSEM}
\title{Reanalyse powerNLSEM object}
\usage{
reanalyse.powerNLSEM(
  out,
  power_modeling_method,
  powerLevels,
  alpha,
  uncertainty_method = ""
)
}
\arguments{
\item{out}{object of class powerNLSEM}

\item{power_modeling_method}{Character indicating the power modeling method used. This is only relevant when plot = "power_model" is used. Default to NULL, indicating to use the same power modeling method as was used in the powerNLSEM function.}

\item{powerLevels}{Power levels for which the desired sample sizes should be computed. Needs to be a vector.}

\item{alpha}{Alpha value used for confidence intervals, when se = TRUE. Default to NULL, indicating to use the same alpha as was used in the powerNLSEM function.}
}
\value{
Returns list of sample sizes per effect. \code{Nall} refers to the sample size required per power level for all coefficients. \code{Npower} is a matrix containing the desired sample sizes per effect for every power level.
}
\description{
Reanalyse powerNLSEM object
}
