% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate data from lavModel object}
\usage{
simulateData(n, lavModel, appendLVs = F)
}
\arguments{
\item{n}{sample size}

\item{lavModel}{lavModel object}

\item{appendLVs}{logical whether latent variables should be observed. Default to FALSE. (For developmental purposes)}
}
\description{
simulate data from lavModel object
}
