% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateNLSEM}
\alias{simulateNLSEM}
\title{simulate data from lavModel object}
\usage{
simulateNLSEM(
  n,
  lavModel,
  appendLVs = F,
  lavModel_attributes = NULL,
  matrices = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{sample size}

\item{lavModel}{lavModel object}

\item{appendLVs}{logical whether latent variables should be observed. Default to FALSE. (For developmental purposes)}

\item{lavModel_attributes}{attributes of the lavModel object. If NULL, this is computed from lavModel. Default to NULL.}

\item{matrices}{computed matrices for simulation. If NULL, this is computed from lavModel and lavModel_attributes. Default to NULL.}
}
\description{
simulate data from lavModel object
}
