% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR.R
\name{SR}
\alias{SR}
\title{Scale Regression approach}
\usage{
SR(lavModel_Analysis, data, data_transformations = NULL)
}
\arguments{
\item{lavModel_Analysis}{the lavModel_Analysis object}

\item{data}{set to fit}

\item{data_transformations}{Data transformations}
}
\description{
Scale Regression approach
}
