% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.powerNLSEM.R
\name{summary.powerNLSEM}
\alias{summary.powerNLSEM}
\title{Summary function for powerNLSEM objects}
\usage{
\method{summary}{powerNLSEM}(object, test = NULL, alpha = NULL, ...)
}
\arguments{
\item{object}{Result of powerNLSEM function estimating the MSPE. \code{object} must be of class \code{"powerNLSEM"}.}

\item{test}{Should the parameter be tested with a directed hypothesis (onesided) or with an undirected hypothesis (twosided, also equivalent to Wald-Test for single parameter). Default to \code{NULL} (if \code{NULL}, \code{test} of the original MSPE is used).}

\item{alpha}{Type I-error rate for significance decision. Default to \code{NULL} (if \code{NULL}, \code{alpha} of the original MSPE is used).}

\item{...}{Further arguments to use in \code{summary}.}
}
\value{
summary of powerNLSEM object
}
\description{
Summary function for powerNLSEM objects
}
