% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.powerNLSEM.R
\name{plot.powerNLSEM}
\alias{plot.powerNLSEM}
\title{plot powerNLSEM object}
\usage{
\method{plot}{powerNLSEM}(out, min_num_bins = 10)
}
\arguments{
\item{out}{object of class powerNLSEM}

\item{min_num_bins}{minimal number of bins used for aggregating results. Default to 10.}
}
\description{
plot powerNLSEM object
}
