% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.powerNLSEM.R
\name{print.powerNLSEM}
\alias{print.powerNLSEM}
\title{print powerNLSEM objects}
\usage{
\method{print}{powerNLSEM}(x, ...)
}
\arguments{
\item{x}{object of class powerNLSEM}
}
\value{
\code{powerNLSEM} object
}
\description{
print powerNLSEM objects
}
