% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPI.R
\name{UPI}
\alias{UPI}
\title{Unconstrained Product Indicator approach by Marsh et al. (2004), with extensions by Kelava and Brandt (2009)}
\usage{
UPI(lavModel_Analysis, data, data_transformations = NULL)
}
\arguments{
\item{lavModel_Analysis}{the lavModel_Analysis object}

\item{data}{set to fit}

\item{data_transformations}{Data transformations}
}
\description{
Unconstrained Product Indicator approach by Marsh et al. (2004), with extensions by Kelava and Brandt (2009)
}
